requirements.txt
apache-libcloud==3.3.1
fasteners==0.16
pyfakefs==4.4.0
pytest==6.2.4

